requirements.txt
jupyter 
matplotlib
numpy
pandas
scikit-learn
seaborn
statsmodels
jupyterlab
nltk
flask
flask_restful
pillow